requirements.txt
colorama
PyYAML
pyjwt
dnspython
